Add in an age restriction to stop recent folders being archived.

+++ zip-to-archive.py
@@ -5,11 +5,24 @@ import os
 from subprocess import run
 from shutil import rmtree
 from time import time
+from datetime import timedelta, datetime
 
 """Zip up each directory in the current folder as separate archives to a specified location (default hardcoded).
   The original directories will be deleted."""
 
 _ZIP_OUTPUT_DIR = r"c:\Test\ZippedOutput"
+_DEFAULT_MIN_DAYS_OLD = 100    # days
+
+
+def calc_birth_date_from_age(days_old: int):
+    """Returns a timestamp that is days_from_today in the past. Use negative value to go forwards."""
+    delta = timedelta(days=days_old)
+    return (datetime.today() - delta).timestamp()
+
+
+def calc_age_from_modtime(modtime: int):
+    delta = datetime.today().timestamp() - modtime/1e9
+    return timedelta(seconds=delta).days
 
 
 def get_latest_mod_date_in_dir(path):
@@ -27,11 +40,12 @@ def get_latest_mod_date_in_dir(path):
         return dir_stat.st_mtime_ns
 
 
-def main(zip_folder):
+def main(zip_folder, min_days_old):
     if not (os.path.exists(zip_folder)):
         print("Error: Output folder '{0}' does not exist.".format(zip_folder))
         return
 
+    cut_off_date_ns = int(calc_birth_date_from_age(min_days_old)*1e9)
     dirs = [d for d in os.listdir(".") if os.path.isdir(d)]
 
     for d in dirs:
@@ -42,19 +56,30 @@ def main(zip_folder):
             print("Error: Archive %s already exists. Skipping." % d)
             continue
 
-        mod_time = get_latest_mod_date_in_dir(d)
-        zip_command = r'7z a "{0}" ".\{1}\*" -sdel'.format(archive_name, d)
+        modtime = get_latest_mod_date_in_dir(d)
+        if modtime > cut_off_date_ns:
+            print("Skipping {}: only {} days old (less than {})."
+                  .format(d, calc_age_from_modtime(modtime), min_days_old))
+            continue
+
+        zip_command = r'7z a "{0}" ".\{1}\*" -sdel -bso0'.format(archive_name, d)
         result = run(zip_command)
 
         if result.returncode == 0 and os.path.exists(archive_name):
-            os.utime(archive_name, ns=(mod_time, mod_time))
+            os.utime(archive_name, ns=(modtime, modtime))
             rmtree(d, ignore_errors=True)
 
 
 if __name__ == "__main__":
     _zip_folder = _ZIP_OUTPUT_DIR
+    _min_days_old = _DEFAULT_MIN_DAYS_OLD
+
     if len(sys.argv) > 1:
         _zip_folder = sys.argv[1]
-        print("zip folder: {}" .format(_zip_folder))
 
-    main(_zip_folder)
+    if len(sys.argv) > 2:
+        _min_days_old = int(sys.argv[2])
+
+    print("zip folder: {}\nmin days old: {}".format(_zip_folder, _min_days_old))
+    main(_zip_folder, _min_days_old)
+    print("All done!")
